+++ rs/bitcoin/adapter/src/connectionmanager.rs
@@ -7,30 +7,30 @@ use std::{
 use bitcoin::p2p::ServiceFlags;
 
 use bitcoin::p2p::{
+    Address, Magic,
     message::{CommandString, NetworkMessage},
     message_network::VersionMessage,
-    Address, Magic,
 };
-use ic_logger::{error, info, trace, warn, ReplicaLogger};
+use ic_logger::{ReplicaLogger, error, info, trace, warn};
 use rand::prelude::*;
 use thiserror::Error;
 use tokio::{
-    sync::mpsc::{channel, Sender},
-    sync::mpsc::{unbounded_channel, Receiver},
+    sync::mpsc::{Receiver, unbounded_channel},
+    sync::mpsc::{Sender, channel},
     task::JoinHandle,
 };
 
 use crate::{
+    Channel, ChannelError, Command, ProcessEvent, ProcessNetworkMessage,
+    ProcessNetworkMessageError,
     addressbook::{
-        validate_services, AddressBook, AddressBookError, AddressEntry, AddressTimestamp,
+        AddressBook, AddressBookError, AddressEntry, AddressTimestamp, validate_services,
     },
     common::{BlockHeight, BlockchainNetwork, DEFAULT_CHANNEL_BUFFER_SIZE, MINIMUM_VERSION_NUMBER},
     config::Config,
     connection::{Connection, ConnectionConfig, ConnectionState, PingState},
     metrics::RouterMetrics,
     stream::{StreamConfig, StreamEvent, StreamEventKind},
-    Channel, ChannelError, Command, ProcessEvent, ProcessNetworkMessage,
-    ProcessNetworkMessageError,
 };
 
 type StreamConfigOf<Network> =
@@ -258,13 +258,12 @@ impl<Network: BlockchainNetwork> ConnectionManager<Network> {
     fn flag_seed_addr_retrieval_timeouts(&mut self) {
         let now = SystemTime::now();
         for conn in self.connections.values_mut() {
-            if let AddressEntry::Seed(_) = *conn.address_entry() {
-                if let ConnectionState::AwaitingAddresses { timestamp } = *conn.state() {
-                    let expires_at =
-                        timestamp + Duration::from_secs(SEED_ADDR_RETRIEVED_TIMEOUT_SECS);
-                    if expires_at <= now {
-                        conn.discard();
-                    }
+            if let AddressEntry::Seed(_) = *conn.address_entry()
+                && let ConnectionState::AwaitingAddresses { timestamp } = *conn.state()
+            {
+                let expires_at = timestamp + Duration::from_secs(SEED_ADDR_RETRIEVED_TIMEOUT_SECS);
+                if expires_at <= now {
+                    conn.discard();
                 }
             }
         }
@@ -360,7 +359,7 @@ impl<Network: BlockchainNetwork> ConnectionManager<Network> {
 
         // The node address that will be receiving this message.
         let receiver = Address::new(addr, ServiceFlags::NETWORK | ServiceFlags::NETWORK_LIMITED);
-        let nonce: u64 = self.rng.gen();
+        let nonce: u64 = self.rng.r#gen();
         let user_agent = String::from(USER_AGENT);
         let message = <NetworkMessageOf<Network>>::Version(VersionMessage::new(
             self.p2p_protocol_version,
@@ -389,7 +388,7 @@ impl<Network: BlockchainNetwork> ConnectionManager<Network> {
 
     /// This function is used to send a `ping` message to a specified connection.
     fn send_ping(&mut self, addr: &SocketAddr) -> ConnectionManagerResult<()> {
-        let nonce = self.rng.gen();
+        let nonce = self.rng.r#gen();
         let conn = self.get_connection(addr)?;
         conn.expect_pong(nonce);
         self.send_to(addr, NetworkMessage::Ping(nonce))
@@ -480,8 +479,7 @@ impl<Network: BlockchainNetwork> ConnectionManager<Network> {
 
         trace!(
             self.logger,
-            "Completed the version handshake with {}",
-            address
+            "Completed the version handshake with {}", address
         );
         Ok(())
     }
@@ -549,10 +547,10 @@ impl<Network: BlockchainNetwork> ConnectionManager<Network> {
             return Err(ProcessNetworkMessageError::InvalidMessage);
         }
 
-        if let Ok(conn) = self.get_connection(address) {
-            if let AddressEntry::Seed(_) = conn.address_entry() {
-                conn.disconnect();
-            }
+        if let Ok(conn) = self.get_connection(address)
+            && let AddressEntry::Seed(_) = conn.address_entry()
+        {
+            conn.disconnect();
         }
 
         if self.address_book.has_enough_addresses() {
@@ -710,7 +708,7 @@ mod test {
     use super::*;
     use crate::config::test::ConfigBuilder;
     use bitcoin::p2p::ServiceFlags;
-    use bitcoin::{block::Header, Block, Network};
+    use bitcoin::{Block, Network, block::Header};
     use ic_logger::replica_logger::no_op_logger;
     use ic_metrics::MetricsRegistry;
     use std::str::FromStr;