+++ rs/boundary_node/ic_boundary/src/routes.rs
@@ -12,9 +12,9 @@ use axum::{
     response::{IntoResponse, Response},
 };
 use candid::{CandidType, Principal};
-use ic_bn_lib::http::{proxy, Client as HttpClient};
+use ic_bn_lib::http::{Client as HttpClient, proxy};
 pub use ic_bn_lib::types::RequestType;
-use ic_types::{messages::ReplicaHealthStatus, CanisterId, SubnetId};
+use ic_types::{CanisterId, SubnetId, messages::ReplicaHealthStatus};
 use serde::Deserialize;
 use url::Url;
 
@@ -257,10 +257,10 @@ pub(crate) mod test {
     use std::sync::Arc;
 
     use anyhow::Error;
-    use axum::{body::Body, http::Request, routing::method_routing::get, Router};
+    use axum::{Router, body::Body, http::Request, routing::method_routing::get};
     use http::{
-        header::{HeaderName, HeaderValue, CONTENT_TYPE, X_CONTENT_TYPE_OPTIONS, X_FRAME_OPTIONS},
         StatusCode,
+        header::{CONTENT_TYPE, HeaderName, HeaderValue, X_CONTENT_TYPE_OPTIONS, X_FRAME_OPTIONS},
     };
     use ic_bn_lib::{
         http::headers::{
@@ -270,22 +270,22 @@ pub(crate) mod test {
         principal,
     };
     use ic_types::{
+        PrincipalId,
         messages::{
             Blob, HttpCallContent, HttpCanisterUpdate, HttpQueryContent, HttpReadState,
             HttpReadStateContent, HttpRequestEnvelope, HttpStatusResponse, HttpUserQuery,
         },
-        PrincipalId,
     };
     use tower::Service;
 
     use crate::{
         http::{
-            handlers::{health, status},
             PATH_HEALTH, PATH_STATUS,
+            handlers::{health, status},
         },
         persist::{Persist, Persister},
         snapshot::test::test_registry_snapshot,
-        test_utils::{setup_test_router, TestHttpClient},
+        test_utils::{TestHttpClient, setup_test_router},
     };
 
     fn assert_header(headers: &http::HeaderMap, name: HeaderName, expected_value: &str) {