+++ rs/ethereum/ledger-suite-orchestrator/src/management/mod.rs
@@ -64,12 +64,12 @@ impl fmt::Display for Reason {
     fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Self::OutOfCycles => write!(fmt, "the canister is out of cycles"),
-            Self::CanisterError(msg) => write!(fmt, "canister error: {}", msg),
+            Self::CanisterError(msg) => write!(fmt, "canister error: {msg}"),
             Self::Rejected(msg) => {
-                write!(fmt, "the management canister rejected the call: {}", msg)
+                write!(fmt, "the management canister rejected the call: {msg}")
             }
-            Reason::TransientInternalError(msg) => write!(fmt, "transient internal error: {}", msg),
-            Reason::InternalError(msg) => write!(fmt, "internal error: {}", msg),
+            Reason::TransientInternalError(msg) => write!(fmt, "transient internal error: {msg}"),
+            Reason::InternalError(msg) => write!(fmt, "internal error: {msg}"),
         }
     }
 }
@@ -84,8 +84,7 @@ impl Reason {
             | RejectionCode::SysFatal
             | RejectionCode::DestinationInvalid
             | RejectionCode::Unknown => Self::InternalError(format!(
-                "rejection code: {:?}, rejection message: {}",
-                reject_code, reject_message
+                "rejection code: {reject_code:?}, rejection message: {reject_message}"
             )),
         }
     }